Failures can happen prior to initializing the Kroxylicious operator.
This then leads to issues debugging the original failure...

+++ kroxylicious-systemtests/src/test/java/io/kroxylicious/systemtests/filters/AuthorizationST.java
@@ -86,7 +86,9 @@ class AuthorizationST extends AbstractST {
 
     @AfterAll
     void cleanUp() {
-        kroxyliciousOperator.delete();
+        if (kroxyliciousOperator != null) {
+            kroxyliciousOperator.delete();
+        }
     }
 
     @Test
